+++ src/gdalraster.cpp
@@ -141,24 +141,6 @@ bool GDALRaster::setGeoTransform(std::vector<double> transform) {
 	}
 }
 
-std::vector<double> GDALRaster::bbox() const {
-	std::vector<double> gt = this->getGeoTransform();
-	double xmin = gt[0];
-	double xmax = xmin + gt[1] * this->getRasterXSize();
-	double ymax = gt[3];
-	double ymin = ymax + gt[5] * this->getRasterYSize();
-	std::vector<double> ret = {xmin, ymin, xmax, ymax};
-	return ret;
-}
-
-std::vector<double> GDALRaster::res() const {
-	std::vector<double> gt = this->getGeoTransform();
-	double pixel_width = gt[1];
-	double pixel_height = std::fabs(gt[5]);
-	std::vector<double> ret = {pixel_width, pixel_height};
-	return ret;
-}
-
 int GDALRaster::getRasterCount() const {
 	if (!this->isOpen())
 		Rcpp::stop("Raster dataset is not open.");
@@ -201,11 +183,45 @@ bool GDALRaster::setProjection(std::string projection) {
 	}
 }
 
+std::vector<double> GDALRaster::bbox() const {
+	if (!this->isOpen())
+		Rcpp::stop("Raster dataset is not open.");
+		
+	std::vector<double> gt = this->getGeoTransform();
+	double xmin = gt[0];
+	double xmax = xmin + gt[1] * this->getRasterXSize();
+	double ymax = gt[3];
+	double ymin = ymax + gt[5] * this->getRasterYSize();
+	std::vector<double> ret = {xmin, ymin, xmax, ymax};
+	return ret;
+}
+
+std::vector<double> GDALRaster::res() const {
+	if (!this->isOpen())
+		Rcpp::stop("Raster dataset is not open.");
+		
+	std::vector<double> gt = this->getGeoTransform();
+	double pixel_width = gt[1];
+	double pixel_height = std::fabs(gt[5]);
+	std::vector<double> ret = {pixel_width, pixel_height};
+	return ret;
+}
+
+std::vector<int> GDALRaster::dim() const {
+	if (!this->isOpen())
+		Rcpp::stop("Raster dataset is not open.");
+		
+	std::vector<int> ret = {this->getRasterXSize(),
+								this->getRasterYSize(),
+								this->getRasterCount()};
+	return ret;
+}
+
 std::vector<int> GDALRaster::getBlockSize(int band) const {
 	if (!this->isOpen())
 		Rcpp::stop("Raster dataset is not open.");
 		
-	GDALRasterBandH hBand = GDALGetRasterBand(hDataset, band);
+	GDALRasterBandH hBand = this->_getBand(band);
 	int nBlockXSize, nBlockYSize;
 	GDALGetBlockSize(hBand, &nBlockXSize, &nBlockYSize);
 	std::vector<int> ret = {nBlockXSize, nBlockYSize};
@@ -216,15 +232,48 @@ int GDALRaster::getOverviewCount(int band) const {
 	if (!this->isOpen())
 		Rcpp::stop("Raster dataset is not open.");
 		
-	GDALRasterBandH hBand = GDALGetRasterBand(hDataset, band);
+	GDALRasterBandH hBand = this->_getBand(band);
 	return GDALGetOverviewCount(hBand);
+}
+
+void GDALRaster::buildOverviews(std::string resampling,
+		std::vector<int> levels, std::vector<int> bands) {
+
+	if (!this->isOpen())
+		Rcpp::stop("Raster dataset is not open.");
+		
+	int nOvr;
+	int* panOvrList = NULL;
+	if (levels.size() == 1 && levels[0] == 0) {
+		nOvr = 0;
+	}
+	else {
+		nOvr = levels.size();
+		panOvrList = levels.data();
+	}
+	
+	int nBands;
+	int* panBandList = NULL;
+	if (bands.size() == 1 && bands[0] == 0) {
+		nBands = 0;
+	}
+	else {
+		nBands = bands.size();
+		panBandList = bands.data();
+	}
+	
+	CPLErr err = GDALBuildOverviews(hDataset, resampling.c_str(), nOvr,
+			panOvrList, nBands, panBandList, GDALTermProgressR, NULL);
+
+	if (err == CE_Failure)
+		Rcpp::stop("Build overviews failed.");
 }
 
 std::string GDALRaster::getDataTypeName(int band) const {
 	if (!this->isOpen())
 		Rcpp::stop("Raster dataset is not open.");
 		
-	GDALRasterBandH hBand = GDALGetRasterBand(hDataset, band);
+	GDALRasterBandH hBand = this->_getBand(band);
 	return GDALGetDataTypeName(GDALGetRasterDataType(hBand));
 }
 
@@ -234,7 +283,7 @@ Rcpp::NumericVector GDALRaster::getStatistics(int band,	bool approx_ok,
 	if (!this->isOpen())
 		Rcpp::stop("Raster dataset is not open.");
 	
-	GDALRasterBandH hBand = GDALGetRasterBand(hDataset, band);
+	GDALRasterBandH hBand = this->_getBand(band);
 	double min, max, mean, sd;
 	CPLErr err;
 	
@@ -265,7 +314,7 @@ bool GDALRaster::hasNoDataValue(int band) const {
 	if (!this->isOpen())
 		Rcpp::stop("Raster dataset is not open.");
 		
-	GDALRasterBandH hBand = GDALGetRasterBand(hDataset, band);
+	GDALRasterBandH hBand = this->_getBand(band);;
 	int hasNoData;
 	GDALGetRasterNoDataValue(hBand, &hasNoData);
 	return hasNoData;
@@ -276,7 +325,7 @@ double GDALRaster::getNoDataValue(int band) const {
 		Rcpp::stop("Raster dataset is not open.");
 		
 	if (this->hasNoDataValue(band)) {
-		GDALRasterBandH hBand = GDALGetRasterBand(hDataset, band);
+		GDALRasterBandH hBand = this->_getBand(band);;
 		return GDALGetRasterNoDataValue(hBand, NULL);
 	}
 	else {
@@ -291,7 +340,7 @@ bool GDALRaster::setNoDataValue(int band, double nodata_value) {
 	if (GDALGetAccess(hDataset) == GA_ReadOnly)
 		Rcpp::stop("Cannot set nodata value (GA_ReadOnly).");
 
-	GDALRasterBandH hBand = GDALGetRasterBand(hDataset, band);
+	GDALRasterBandH hBand = this->_getBand(band);;
 	if (GDALSetRasterNoDataValue(hBand, nodata_value) == CE_Failure) {
 		Rcpp::Rcerr << "Set nodata value failed.\n";
 		return false;
@@ -308,7 +357,7 @@ void GDALRaster::deleteNoDataValue(int band) {
 	if (GDALGetAccess(hDataset) == GA_ReadOnly)
 		Rcpp::stop("Cannot delete nodata value (GA_ReadOnly).");
 
-	GDALRasterBandH hBand = GDALGetRasterBand(hDataset, band);
+	GDALRasterBandH hBand = this->_getBand(band);;
 	if (GDALDeleteRasterNoDataValue(hBand) == CE_Failure) {
 		Rcpp::stop("Delete nodata value failed.");
 	}
@@ -318,15 +367,32 @@ std::string GDALRaster::getUnitType(int band) const {
 	if (!this->isOpen())
 		Rcpp::stop("Raster dataset is not open.");
 		
-	GDALRasterBandH hBand = GDALGetRasterBand(hDataset, band);
+	GDALRasterBandH hBand = this->_getBand(band);;
 	return GDALGetRasterUnitType(hBand);
+}
+
+bool GDALRaster::setUnitType(int band, std::string unit_type) {
+	if (!this->isOpen())
+		Rcpp::stop("Raster dataset is not open.");
+		
+	if (GDALGetAccess(hDataset) == GA_ReadOnly)
+		Rcpp::stop("Cannot set unit type (GA_ReadOnly).");
+
+	GDALRasterBandH hBand = this->_getBand(band);;
+	if (GDALSetRasterUnitType(hBand, unit_type.c_str()) == CE_Failure) {
+		Rcpp::Rcerr << "Set unit type failed.\n";
+		return false;
+	}
+	else {
+		return true;
+	}
 }
 
 bool GDALRaster::hasScale(int band) const {
 	if (!this->isOpen())
 		Rcpp::stop("Raster dataset is not open.");
 		
-	GDALRasterBandH hBand = GDALGetRasterBand(hDataset, band);
+	GDALRasterBandH hBand = this->_getBand(band);;
 	int hasScale;
 	GDALGetRasterScale(hBand, &hasScale);
 	return hasScale;
@@ -337,19 +403,36 @@ double GDALRaster::getScale(int band) const {
 		Rcpp::stop("Raster dataset is not open.");
 		
 	if (this->hasScale(band)) {
-		GDALRasterBandH hBand = GDALGetRasterBand(hDataset, band);
+		GDALRasterBandH hBand = this->_getBand(band);;
 		return GDALGetRasterScale(hBand, NULL);
 	}
 	else {
 		return NA_REAL;
 	}
-}	
+}
+
+bool GDALRaster::setScale(int band, double scale) {
+	if (!this->isOpen())
+		Rcpp::stop("Raster dataset is not open.");
+		
+	if (GDALGetAccess(hDataset) == GA_ReadOnly)
+		Rcpp::stop("Cannot set scale (GA_ReadOnly).");
+
+	GDALRasterBandH hBand = this->_getBand(band);;
+	if (GDALSetRasterScale(hBand, scale) == CE_Failure) {
+		Rcpp::Rcerr << "Set scale failed.\n";
+		return false;
+	}
+	else {
+		return true;
+	}
+}
 
 bool GDALRaster::hasOffset(int band) const {
 	if (!this->isOpen())
 		Rcpp::stop("Raster dataset is not open.");
 		
-	GDALRasterBandH hBand = GDALGetRasterBand(hDataset, band);
+	GDALRasterBandH hBand = this->_getBand(band);;
 	int hasOffset;
 	GDALGetRasterOffset(hBand, &hasOffset);
 	return hasOffset;
@@ -360,11 +443,28 @@ double GDALRaster::getOffset(int band) const {
 		Rcpp::stop("Raster dataset is not open.");
 		
 	if (this->hasOffset(band)) {
-		GDALRasterBandH hBand = GDALGetRasterBand(hDataset, band);
+		GDALRasterBandH hBand = this->_getBand(band);;
 		return GDALGetRasterOffset(hBand, NULL);
 	}
 	else {
 		return NA_REAL;
+	}
+}
+
+bool GDALRaster::setOffset(int band, double offset) {
+	if (!this->isOpen())
+		Rcpp::stop("Raster dataset is not open.");
+		
+	if (GDALGetAccess(hDataset) == GA_ReadOnly)
+		Rcpp::stop("Cannot set offset (GA_ReadOnly).");
+
+	GDALRasterBandH hBand = this->_getBand(band);;
+	if (GDALSetRasterOffset(hBand, offset) == CE_Failure) {
+		Rcpp::Rcerr << "Set offset failed.\n";
+		return false;
+	}
+	else {
+		return true;
 	}
 }
 
@@ -383,7 +483,7 @@ Rcpp::CharacterVector GDALRaster::getMetadata(int band,
 			papszMetadata = GDALGetMetadata(hDataset, domain.c_str());
 	}
 	else {
-		GDALRasterBandH hBand = GDALGetRasterBand(hDataset, band);
+		GDALRasterBandH hBand = this->_getBand(band);;
 		if (domain == "")
 			papszMetadata = GDALGetMetadata(hBand, NULL);
 		else
@@ -421,7 +521,7 @@ std::string GDALRaster::getMetadataItem(int band, std::string mdi_name,
 					GDALGetMetadataItem(hDataset, mdi_name.c_str(), NULL) );
 	}
 	else {
-		GDALRasterBandH hBand = GDALGetRasterBand(hDataset, band);
+		GDALRasterBandH hBand = this->_getBand(band);;
 		if (GDALGetMetadataItem(hBand, mdi_name.c_str(), domain_) != NULL) 
 			mdi += std::string(
 					GDALGetMetadataItem(hBand, mdi_name.c_str(), domain_) );
@@ -430,13 +530,39 @@ std::string GDALRaster::getMetadataItem(int band, std::string mdi_name,
 	return mdi;
 }
 
+void GDALRaster::setMetadataItem(int band, std::string mdi_name, 
+		std::string mdi_value, std::string domain) {
+	
+	if (!this->isOpen())
+		Rcpp::stop("Raster dataset is not open.");
+		
+	if (GDALGetAccess(hDataset) == GA_ReadOnly)
+		Rcpp::stop("Cannot set metadata item (GA_ReadOnly).");
+
+	const char* domain_ = NULL;
+	if (domain != "")
+		domain_ = domain.c_str();
+
+	if (band == 0) {
+		if (GDALSetMetadataItem(hDataset, mdi_name.c_str(), mdi_value.c_str(),
+								domain_) != CE_None)
+			Rcpp::stop("Set metadata item failed.");
+	}
+	else {
+		GDALRasterBandH hBand = this->_getBand(band);;
+		if (GDALSetMetadataItem(hBand, mdi_name.c_str(), mdi_value.c_str(),
+								domain_) != CE_None)
+			Rcpp::stop("Set metadata item failed.");
+	}
+}
+
 SEXP GDALRaster::read(int band, int xoff, int yoff, int xsize, int ysize,
 		int out_xsize, int out_ysize) const {
 	
 	if (!this->isOpen())
 		Rcpp::stop("Raster dataset is not open.");
 	
-	GDALRasterBandH hBand = GDALGetRasterBand(hDataset, band);
+	GDALRasterBandH hBand = this->_getBand(band);;
 	GDALDataType eDT = GDALGetRasterDataType(hBand);
 	
 	CPLErr err;
@@ -549,7 +675,7 @@ void GDALRaster::write(int band, int xoff, int yoff, int xsize, int ysize,
 		// real data types
 	
 		eBufType = GDT_Float64;
-		GDALRasterBandH hBand = GDALGetRasterBand(hDataset, band);
+		GDALRasterBandH hBand = this->_getBand(band);;
 		std::vector<double> buf_ = Rcpp::as<std::vector<double>>(rasterData);
 		if (buf_.size() != ((std::size_t) (xsize * ysize)))
 			Rcpp::stop("Size of input data is not the same as region size.");
@@ -561,7 +687,7 @@ void GDALRaster::write(int band, int xoff, int yoff, int xsize, int ysize,
 		// complex data types
 	
 		eBufType = GDT_CFloat64;
-		GDALRasterBandH hBand = GDALGetRasterBand(hDataset, band);
+		GDALRasterBandH hBand = this->_getBand(band);;
 		std::vector<std::complex<double>> buf_ = 
 			Rcpp::as<std::vector<std::complex<double>>>(rasterData);
 		if (buf_.size() != ((std::size_t) (xsize * ysize)))
@@ -584,35 +710,20 @@ void GDALRaster::fillRaster(int band, double value, double ivalue) {
 	if (GDALGetAccess(hDataset) == GA_ReadOnly)
 		Rcpp::stop("Dataset is read-only.");
 	
-	GDALRasterBandH hBand = GDALGetRasterBand(hDataset, band);
+	GDALRasterBandH hBand = this->_getBand(band);;
 	if (GDALFillRaster(hBand, value, ivalue) == CE_Failure) {
 		Rcpp::stop("Fill raster failed.");
 	}
 }
 
-int GDALRaster::getChecksum(int band, int xoff, int yoff, int xsize, int ysize) const {
-	if (!this->isOpen())
-		Rcpp::stop("Raster dataset is not open.");
-	
-	GDALRasterBandH hBand = GDALGetRasterBand(hDataset, band);
-	return GDALChecksumImage(hBand, xoff, yoff, xsize, ysize);
-}
+int GDALRaster::getChecksum(int band, int xoff, int yoff,
+		int xsize, int ysize) const {
 
-void GDALRaster::_setMetadataItem(int band, std::string mdi_name, 
-		std::string mdi_value, std::string domain) {
-		
-	// currently only for GDALRasterBand metadata item
-	
 	if (!this->isOpen())
 		Rcpp::stop("Raster dataset is not open.");
-		
-	if (GDALGetAccess(hDataset) == GA_ReadOnly)
-		Rcpp::stop("Cannot set metadata item (GA_ReadOnly).");
-		
-	GDALRasterBandH hBand = GDALGetRasterBand(hDataset, band);
-	if (GDALSetMetadataItem(hBand, mdi_name.c_str(), mdi_value.c_str(),
-							domain.c_str()) != CE_None)
-		Rcpp::stop("Set metadata item failed.");
+	
+	GDALRasterBandH hBand = this->_getBand(band);;
+	return GDALChecksumImage(hBand, xoff, yoff, xsize, ysize);
 }
 
 void GDALRaster::close() {
@@ -620,10 +731,21 @@ void GDALRaster::close() {
 	hDataset = NULL;
 }
 
+// class methods for internal use not exposed in R
+
+GDALRasterBandH GDALRaster::_getBand(int band) const {
+	GDALRasterBandH hBand = GDALGetRasterBand(hDataset, band);
+	if (hBand == NULL)
+		Rcpp::stop("Failed to access the requested band.");
+	return hBand;
+}
+
 RCPP_MODULE(mod_GDALRaster) {
 
     Rcpp::class_<GDALRaster>("GDALRaster")
 
+    .constructor
+    	("Default constructor, no dataset opened.")
     .constructor<std::string, bool>
     	("Usage: new(GDALRaster, filename, read_only=TRUE")
     
@@ -648,20 +770,24 @@ RCPP_MODULE(mod_GDALRaster) {
     	"Return the affine transformation coefficients.")
     .method("setGeoTransform", &GDALRaster::setGeoTransform, 
     	"Set the affine transformation coefficients for this dataset.")
-    .const_method("bbox", &GDALRaster::bbox, 
-    	"Return the bounding box (xmin, ymin, xmax, ymax).")
-    .const_method("res", &GDALRaster::res, 
-    	"Return the resolution (pixel width, pixel height).")
     .const_method("getRasterCount", &GDALRaster::getRasterCount, 
     	"Return the number of raster bands on this dataset.")
     .const_method("getProjectionRef", &GDALRaster::getProjectionRef, 
     	"Return the projection definition for this dataset.")
     .method("setProjection", &GDALRaster::setProjection, 
     	"Set the projection reference string for this dataset.")
+    .const_method("bbox", &GDALRaster::bbox, 
+    	"Return the bounding box (xmin, ymin, xmax, ymax).")
+    .const_method("res", &GDALRaster::res, 
+    	"Return the resolution (pixel width, pixel height).")
+    .const_method("dim", &GDALRaster::dim, 
+    	"Return raster dimensions (xsize, ysize, number of bands).")
     .const_method("getBlockSize", &GDALRaster::getBlockSize, 
     	"Get the natural block size of this band.")
     .const_method("getOverviewCount", &GDALRaster::getOverviewCount, 
     	"Return the number of overview layers available.")
+    .method("buildOverviews", &GDALRaster::buildOverviews, 
+    	"Build raster overview(s).")
     .const_method("getDataTypeName", &GDALRaster::getDataTypeName, 
     	"Get name of the data type for this band.")
     .const_method("getStatistics", &GDALRaster::getStatistics, 
@@ -674,14 +800,22 @@ RCPP_MODULE(mod_GDALRaster) {
     	"Delete the nodata value for this band.")
     .const_method("getUnitType", &GDALRaster::getUnitType, 
     	"Get name of the raster value units (e.g., m or ft).")
+    .method("setUnitType", &GDALRaster::setUnitType, 
+    	"Set name of the raster value units (e.g., m or ft).")
     .const_method("getScale", &GDALRaster::getScale, 
-    	"Return the raster value scale.")
+    	"Return the raster value scaling ratio.")
+    .method("setScale", &GDALRaster::setScale, 
+    	"Set the raster value scaling ratio.")
     .const_method("getOffset", &GDALRaster::getOffset, 
     	"Return the raster value offset.")
+    .method("setOffset", &GDALRaster::setOffset, 
+    	"Set the raster value offset.")
     .const_method("getMetadata", &GDALRaster::getMetadata, 
     	"Return a list of metadata item=value for a domain.")
     .const_method("getMetadataItem", &GDALRaster::getMetadataItem, 
     	"Return the value of a metadata item.")
+    .method("setMetadataItem", &GDALRaster::setMetadataItem, 
+    	"Set metadata item name=value in domain.")
     .const_method("read", &GDALRaster::read, 
     	"Read a region of raster data for a band.")
     .method("write", &GDALRaster::write, 
@@ -690,8 +824,6 @@ RCPP_MODULE(mod_GDALRaster) {
     	"Fill this band with a constant value.")
     .const_method("getChecksum", &GDALRaster::getChecksum, 
     	"Compute checksum for raster region.")
-    .method(".setMetadataItem", &GDALRaster::_setMetadataItem, 
-    	"Set metadata item name=value in domain.")
     .method("close", &GDALRaster::close, 
     	"Close the GDAL dataset for proper cleanup.")